# api/openapi.yaml
openapi: "3.0.3"
info:
  title: CMDB Platform API
  version: "1.0.0"
  description: Configuration Management Database Platform API
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /assets/{id}/upgrade-recommendations:
    get:
      operationId: getAssetUpgradeRecommendations
      tags: [assets]
      summary: Get upgrade recommendations for an asset
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Upgrade recommendations
  /assets/{id}/upgrade-recommendations/{category}/accept:
    post:
      operationId: acceptUpgradeRecommendation
      tags: [assets]
      summary: Accept an upgrade recommendation
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendation accepted

  # ── Maintenance ─────────────────────────────────────────────────────────
...
  /prediction/rul/{id}:
    get:
      operationId: getAssetRUL
      tags: [prediction]
      summary: Get remaining useful life for an asset
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: RUL data
...
  /upgrade-rules:
    get:
      operationId: getUpgradeRules
      tags: [prediction]
      summary: List upgrade rules for the current tenant
      responses:
        '200':
          description: Upgrade rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      rules:
                        type: array
                        items: { $ref: '#/components/schemas/UpgradeRule' }
                    required: [rules]
                  meta: { $ref: '#/components/schemas/Meta' }
    post:
      operationId: createUpgradeRule
      tags: [prediction]
      summary: Create a new upgrade rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUpgradeRuleRequest' }
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                    required: [id]
                  meta: { $ref: '#/components/schemas/Meta' }
        '400':
          $ref: '#/components/responses/BadRequest'
  /upgrade-rules/{id}:
    put:
      operationId: updateUpgradeRule
      tags: [prediction]
      summary: Update an existing upgrade rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUpgradeRuleRequest' }
      responses:
        '200':
          description: Rule updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      updated: { type: boolean }
                    required: [updated]
                  meta: { $ref: '#/components/schemas/Meta' }
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      operationId: deleteUpgradeRule
      tags: [prediction]
      summary: Delete an upgrade rule
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Rule deleted

  # ── User roles assignment ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
      required: [request_id]
...
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
      required: [error, meta]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UpgradeRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        asset_type: { type: string }
        category: { type: string }
        metric_name: { type: string }
        threshold: { type: number, format: double }
        duration_days: { type: integer }
        priority: { type: string, enum: [low, medium, high] }
        recommendation: { type: string }
        enabled: { type: boolean }
      required: [id, asset_type, category, priority, enabled]
    CreateUpgradeRuleRequest:
      type: object
      properties:
        asset_type: { type: string }
        category: { type: string }
        metric_name: { type: string }
        threshold: { type: number, format: double }
        duration_days: { type: integer, default: 7 }
        priority: { type: string, enum: [low, medium, high], default: medium }
        recommendation: { type: string }
        enabled: { type: boolean, default: true }
      required: [asset_type, category]
    UpdateUpgradeRuleRequest:
      type: object
      properties:
        threshold: { type: number, format: double }
        duration_days: { type: integer }
        priority: { type: string, enum: [low, medium, high] }
        recommendation: { type: string }
        enabled: { type: boolean }

    # ── Inventory scan ─────────────────────────────────────────────────────
...